requirements.txt
cppimport>=22.8.2,<23
docopt>=0.6.2,<0.7
matplotlib>=3.8.2,<4
mpld3>=0.5.10,<0.6
pandas>=2.2.0,<3
pybind11>=2.11.1,<3
pydocstyle>=6.3.0,<7
pyglet==1.5.15
pylint>=3.0.3,<4
setuptools>=80.9.0
numpy>=1.18.0,<2.0
cloudpickle>=1.2.0
torch>=2.0.0
intel-openmp>=2021.4.0
python-dotenv
